fix: file_manager_ensure_dir tried mkdir on file path instead of parent dir
Root cause: ensure_dir('/storage/www/index.html') tried mkdir('index.html')
instead of ensuring '/storage/www/' exists. Now correctly strips
filename and only creates directory components.

+++ main/web_server/handlers/upload_handler.c
@@ -241,7 +241,7 @@ static esp_err_t api_upload_handler(httpd_req_t *req)
 
         size_t written = fwrite(buf, 1, received, file);
         if (written != (size_t)received) {
-            ESP_LOGE(TAG, "fwrite failed: wrote %zu of %d", written, received);
+            ESP_LOGE(TAG, "fwrite failed: wrote %lu of %d", (unsigned long)written, received);
             ret = ESP_FAIL;
             break;
         }
